requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
discord.py==1.5.1
dnspython==2.0.0
idna==2.10
lxml==4.6.1
multidict==4.7.6
pymongo==3.11.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-engineio==3.13.2
python-socketio==4.6.0
python-utils==2.4.0
pytz==2021.1
requests==2.24.0
requests-toolbelt==0.9.1
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.11
websockets==8.1
yarl==1.5.1